requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
loguru==0.7.0
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
PySocks==1.7.1
pytest==7.3.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
